@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "npm run server | ng serve",
+    "start": "ng serve",
     "build": "ng build",
     "server": "node server.js",
     "test": "ng test",
@@ -14,61 +14,73 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^5.2.1",
+    "@angular/cdk": "^5.1.1",
     "@angular/common": "^5.2.1",
     "@angular/compiler": "^5.2.1",
     "@angular/core": "^5.2.1",
     "@angular/forms": "^5.2.1",
     "@angular/http": "^5.2.1",
+    "@angular/material": "^5.1.1",
     "@angular/platform-browser": "^5.2.1",
     "@angular/platform-browser-dynamic": "^5.2.1",
     "@angular/router": "^5.2.1",
+    "@ngrx/effects": "^5.0.1",
+    "@ngrx/entity": "^5.0.1",
+    "@ngrx/router-store": "^5.0.1",
+    "@ngrx/store": "^5.0.0",
+    "@ngrx/store-devtools": "^5.0.1",
     "awesome-typescript-loader": "^3.4.1",
     "bootstrap": "^4.0.0",
     "bootstrap-loader": "^2.2.0",
     "core-js": "^2.4.1",
     "debug": "^3.1.0",
     "dotenv": "^4.0.0",
-    "jquery": "^3.2.1",
-    "popper.js": "^1.12.9",
+    "jquery": "^3.3.1",
+    "popper.js": "^1.13.0",
     "rxjs": "^5.5.2",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular-devkit/schematics": "0.0.46",
-    "@angular/cli": "^1.6.4",
-    "@angular/compiler-cli": "^5.2.1",
-    "@angular/language-service": "^5.2.1",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.9",
+    "@angular/cli": "^1.6.7",
+    "@angular/compiler-cli": "^5.2.3",
+    "@angular/language-service": "^5.2.3",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^4.1.1",
-    "@ngrx/router-store": "^4.1.1",
-    "@ngrx/store": "^4.1.1",
-    "@ngrx/store-devtools": "^4.1.1",
+    "@ngrx/effects": "^5.0.1",
+    "@ngrx/router-store": "^5.0.1",
+    "@ngrx/store": "^5.0.0",
+    "@ngrx/store-devtools": "^5.0.1",
     "@types/hammerjs": "^2.0.35",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.96",
-    "@types/webpack": "^3.8.3",
+    "@types/webpack": "^3.8.4",
     "awesome-typescript-loader": "^3.4.1",
     "bootstrap-loader": "^2.2.0",
     "bootstrap-sass": "^3.3.7",
     "codelyzer": "^4.0.2",
-    "compression-webpack-plugin": "^1.1.3",
+    "compression-webpack-plugin": "^1.1.6",
     "copy-webpack-plugin": "^4.3.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.4.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "lowdb": "^1.0.0",
+    "morgan": "^1.9.0",
+    "node-uuid": "^1.4.8",
+    "path": "^0.12.7",
     "protractor": "~5.1.2",
     "resolve-url-loader": "^2.2.1",
     "ts-node": "^3.3.0",
     "tslint": "~5.7.0",
-    "typescript": "^2.5.3",
+    "typescript": "^2.6.2",
     "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.11.0"
+    "webpack-dev-server": "^2.11.0",
+    "winston": "2.4.0"
   }
 }
